% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Create a local copy of example data}
\usage{
get_data()
}
\value{
logical TRUE if the operation was successful
}
\description{
This function will attempt to create a local copy of the example data
so you can practice reading in data in a loop
}
