% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrProgramming-package.R
\docType{package}
\name{jrProgramming-package}
\alias{jrProgramming-package}
\alias{jrProgramming}
\title{R courses from Jumping RIvers}
\description{
Functions and datasets used in courses by Jumping Rivers.
}
\keyword{package}
